Additions and Modifications to Ship.java
I made a bunch of changes, including the deletion of a whole bunch of
variables and methods. That was done to use a much simpler rotation
system, because the one I had conceived earlier was impractical.
The second change I made was to make way for a health system. This will,
in the future, allow for tracking such that I can delete ships that are
"destroyed" in the game.