Health System Update and ControlSystem Interface
For the health system, I created a maxHealth variable, which sets a
standard limit to the health statistic. Since it is a final variable, I
had to modify the constructor for the Ship class to allow this.
The ControlSystem interface is used simply to make it easier on me when
I start writing code for Entity subclasses that need guidance systems.
Since I'll need code from the Ship class for classes that aren't its
subclass, I decided to make an interface to share some of that code. For
example, I want to create a subclass of Ship that counts for all of the
projectile weapons that I plan to add to the game. Since a subclass of
that will be a guided projectile, I will be able to use the
ControlSystem interface to implement a series of methods that function
similarly to the Ship class.