Physics Engine Additions
This update improves the current physics engine to include the start of
a collision physics system. The CycleRunner class and EntityList class
were modified to make use of this. It also created a class to store some
physics constants, which at the moment include the speed of light and
Isaac Newton's Gravitational Constant.
The update also includes a Projectile class, which will be used as a
superclass for various weapons in the game. I also created an equation
for setting ship health, which if required can be changed or replaced at
any time.