Collision Physics Changes
I modified the Ship-to-Ship collision physics so that they will break
out of their warp jumps if they collide. Also, I added a method so that
if an object hits a miscellaneous piece of debris, it will also break
warp. I also added a getMass() method because it might be useful. The
breaking of warp simplifies things, making it unnecessary to hypothesize
what would happen if a Ship in warp hit something.