Projectiles Now Deal Damage
These changes cause Projectile objects that collide with Ship objects to
cause damage to Ship objects. They will also be destroyed after the
collision is underwent. I also created an official method for dealing
damage to Ship objects, so that it doesn't look as chaotic when I'm
looking for the lines of code that cause Ship objects to take damage.