Modified Mass Calculations
Calculations for the mass of a Ship object are calculated based on the
radius of the Ship, which influences the amount of material needed to
build it, and therefore its mass. Also, it takes into effect the weights
of the weapons being stored. Some changes are also made to compensate
for how slowly Ship objects will accelerate and rotate.