Fixed Missile Code
These changes fix a big problem with the Missile class. The problem was
that Missiles simply did not work. However, I made a series of changes
to fix this problem, as well as a few others. The next change that I
need to make is to have the Missile objects attempt to hit a point in
front of the target, as the missiles seem to orbit the target endlessly.