Added Ship Regeneration and Shielding
Ship objects now have shields that recover, as well as a limited healing
ability that allows the Ship to heal a total of health equal to the
maximum health. The amount of health it has left to recover is stored in
a variable, so that when it reaches zero, it can no longer heal.