Moved All Files
Placed all of the code into a folder named "Engine", which will hold the
game's engine just in case I need it for something else. Any updates to
the engine will be put into this new folder. The final game will be
stored in folders that have yet to be created. Don't be fooled by the
large amount of changes; absolutely nothing is different.