GameMode Changes
I made a bunch of modifications. The first one was to have Ships spawn
in when the game starts. When starting up, it will spawn in a formation
of Ship objects for each team. I also moved MapGenerator to the correct
folder and set it up so that it will create a list of Ship objects to
match the map. I also made a couple of methods in CycleRunner static so
that they would be executable.