Added Entity Colors and Shapes
There are now special shapes for Ships and Structures. Structures other
than WarpGates will be diamonds, Corvettes are triangles, Frigates are
squares, Destroyers are pentagons, and cruisers are hexagons. Also, I've
added colors so that blue represents one's own troops, green is allied,
red is an enemy, yellow is neutral, and white represents not being able
to have a team.