Modified Ship Class Plus Slight Interface Change
I made it so that the two ToggleButton objects are used to toggle
settings for all of the Ship objects. Also, I changed Order parsing so
that it removes any Orders that repeat the same type. The final change
that I added at the last second was to have text display disabled on
inactive Buttons, and to have inactive (as in not accepting input, not
the toggle state) ToggleButton buttons display as off and without text.