Interface Improvements
Code has been written so that the currently added Buttons on the
GameMenu will be usable. Also, I fixed some display code for Entities,
so that they appear in the right spot. Finally, I'm temporarily having
Entity objects display as spheres again because of a problem with
creating the polygons used to represent them.