Server and Client Connection Improvements
I have made improvements that offer several things. First, I fixed the
EntityListWrapper, because it wasn't placed properly, and was causing
problems as a result. Also, I improved the code to the point where I was
able to successfully create an instance where the game actually
connected, but I wasn't able to control the Ships.