Changed Miscellaneous Lines
Just about everything I did affects absolutely nothing. I'm doing this
to isolate a new series of changes that I can work on while I wait for a
computer that can run fast enough to work on the server-client stuff
(the laptop is REALLY slow; Defraggler says "ERROR" for the hard drive's
health).