Partially Implemented New "Perk" System
Rather than the bland Ship modifiers I had implemented earlier, I want
to use a system where a user can choose from a broad selection of
"perks" that only have a positive effect on their Ships. A list of the
perks included in the final version is listed in a javadoc comment above
the addPerk() method.